Adds a CommandPool class which allows to avoid any allocations during operation, acting as a memory pool.
Splits state of the FPU grid and state of the
FPU driver into different enumerations, because
these are different things.

Changes the time-out count to 'unsigned' type
so that incrementing the count cannot cause
undefined behavior, and it becomes easy to
note any change in the count when waiting for
commands to terminate.

Overhaul of the sloppy matching of states
when waiting for command completion. The
new concept maintains an array of counters
which match the sum of the states of each
FPUs. That array is evaluated to a kind
of "minimal operational denominator", so
that failures of FPUs to normally progress
in operation result in "retarded" states.
The summed state results in a bitmask entry.

When waiting for a command, the caller
can specify a bitmask which describes state
events which are desired to wait for.