Add a condition variable to fetching of a new unused command buffer, so that the fetch blocks when no buffers are avilable, rather than returning a null pointer.
This blocking should almost never happen, as we have more command
buffers than are used in normal circumstances, but it is not
impossible that a lot of commands are sent and the tx thread does not
makes much progress with returning the buffers.  In this case, the
provideInstance() method will now block and wait until buffers are
avilable again.