Change the loop counter for filtering configuration messages.
The changed loop counted through the individual wave table list
for each FPU, erasing the entries which already were sent
successfully. However erasing an element means that the
next element will have the same index. This does not work if
the counter is counting updwards, so the direction was
changed to downwards. With counting downwards, the sliding down
of elements will not make a problem.