Fix edge cases in timing computations.
TIME_T_MAX is handled as a special case similar to the floating point
"inf" symbol. This is because of the meaning associated in the
time-out computations which means to "never time out".  Under/Overflow
which does not involve TIME_T_MAX is checked and handled with an
assertion because that really shouldn't happen.