Fix concurrency bug in count of pending commands.
The error in this bug was to assume that the response
to a command would always be processed after the
request was registered as a pending command. However
while the response cannot arrive before the request
was sent in physical time, it can happen that
the registration of the request is delayed -
that means that between the registration of
the request, and processing the response, there
is no warranted before-after relationship.

This turned into an error when decrementing the
number of pending requests, which was written
with the idea in mind that a request could supersede
an earlier request which had not yet timed out:
The number of requests was incremented only if
for that FPU there was no already ongoing command,
and decremented only if a command had been registered.
Given the missing before-after relationship,
this is wrong, at least without making sure that
sent and received requests are synchronized.

The fix consists in letting the number of pending
requests float freely, also assuming negative
values. In addition, the number of unsent commands
was made a property of the FPUArray class, so that
the waitForStatus command can wait for the event of that
number going to zero. The corresponding event
was added.