Use eventfd to notify about new commands in queue.
Previously, the queue was polled and if it was empty, the reaction
was to wait a bit before polling it again. This meant it was either
necessary to poll in a busy loop, or new commands would be delayed,
both of which was undesirable. The change uses an event file descriptor
(eventfd) to notify the tx thread about new commands. This makes
responses to new commands much faster.
Also, another eventfd is used to check for the event of closing the
connection. This allows it to use longer time-outs, which is especially
useful for those CAN commands which can take a long time to complete.