Fix book-keeping of time-outs which caused severe performance problems.
Core of the problem was that the previous approach of storing a
12-byte time-out value in one array per fPU with one index for for
each command code caused the fpu_state structure to be far too big.

This caused lock contention and had the result that the different
threads were fighting for the lock.  This resulted in times as long as
40 seconds for sending 2000 messages to 1000 FPUs. The issue
disappeared when the number of commands (NUM_CAN_COMMANDS) was reduced
to about 6.  The solution is to use a simple list with pending
time-outs with up to four entries. Also, the state structure has been
packed and optimized a bit for space.

A remaining but minor problem at this point is that the caching of
time-out values does not work yet reliably, which requires to search
the full list on every call of RX poll. That takes around 100 usec -
not ideal, but not too bad for now.