Add wait time parameter to waitForState.
This is required because we want to cancel movements
from the Python command line. This requires checking
for interrupts / signal and returning control to the
Python interpreter to check periodically for signals.
Doing this using sleep() however is very inefficient and
would delay responses, so we include a max wait time here.