Correct sending of broadcast commands to send one command to each individual bus.
This change affects only broadcast messages which do not use the EtherCAN gateway SYNC
message. It fixes an error which showed up when broadcast messages were
send to a simulated FPU grid where more than one bus was connected
to one (simulated) gateway. In this case, the message was sent only
to the first CAN bus, but time-outs were registered for every FPU.
This had the result that commands such as executeMotion timed out with missing
responses.

The fix was to send a message to each CAN bus. For this, the methods
for sending a broadcast command were refactored to avoid
excessive duplication. Also, bounds checking of FPU ids in
the pending set needed to be corrected.