Add checks for changed time-out count
The driver methods should return error codes when a time-out occurs.
The time-out is captured by the code in GatewayDriver, which
increases the time-out count for the corresponding FPU
and for the whole grid. The change in the counter
needs to be checked for.

Note that the counter is an unsigned value because
unsigned overflow is fine in C / C++ while signed
overflow causes undefined behavior.

Also, the return code for the time-out had to be differentiated
from a normal pass of the waiting time in waitForState.
Passing that waiting time is not an error because
the user of the driver API might just wait for a short
time if anything happens, e.g. to display a real-time
position display.Therefore, an additional error code
was added, a corresponding wrapped python exception,
and the naming was changed to indicate the different
meaning of the two code.