Add test to EtherCAN simulator for independent datum operations
Extends the EtherCAN simulator and firmware simulation to
independent selection of alpha and beta arm for partial
datum operation.

This option is only used in the verification rig, and works
by setting additional flags in the CAN commands. It needs
a modification to the FPU firmware of protocol V1. The protocol
remains compatible to the normal datum command. To select
the modified protocol, add a command line switch "-V1.1" .