Include error handling for ER_DATUM_TO and ER_DATUM_LIMIT.
The first error code is generated by the FPU when the alpha arm is on
the datum and limit switch, to improve safety somewhat.

The second error is generated when a datum command hits
a time-out which is set in the FPU firmware. In this
case, the driver and simulator sets the FPU state
to ABORTED.

Also fixed a nasty mutation error in the place where the
current position intervals are stored before a
moving command is tried. The error was to store
object references instead of copies, and the
objects were mutated, leading to inconsistent state
when the state was restored.