Fix: Bug in range checking of non-monotonous reversed movements.
A bug in the software protection layer was fixed which affected
reversing movements with the reverseMotion() command when the movement
was not monotonous, but going forwards and backwards. The cause for
the bug was that the stored waveform was, correctly, checked with the
reverse sign, but, incorrectly, with the temporal order of the forward
movement. This had the effect that waveforms were rejected which were
actually safe.