Add faster waveform upload options to the driver and documentation.
Note, these options are not yet fully functional. The reason is that
for the waitForState method, discovery of a finished command is
that its pending bit is cleared. This does not work for
more than one command of the same type at the same time - except
configMotion, all commands are designed to be sent only
after the previous command was confirmed.