Fix bug where waveform generation does not match protocol
The current FPU firmware (1.4.4) does not allow for small
step counts except in the very last segment of a waveform.
This change fixes the waveform genertation so that waveforms
where the alpha and the beta arm have different number of
sections with movement, the shorter movement is padded in the
beginning, rather than on both ends.