revise update of ping_ok flag
The ping_ok flag means that the present step counter values
are not necessarily valid and should be update via a ping.
Because in protocol v2, most responses report the current
positions, it is only rarely necessary to update these flags.
Cases where this is indicated are collisions, limit switch
breaches, and communication errors.