update waveforms rule sets to allow constant acceleration
Make ruleset version V5 the default (allowing for
constant acceleration, as used in the verification rig).

This increases the acceleration allowed for low motor speeds, and
should increase manoeuvrability of the FPUs, to assist more efficient
path planning.

Note - the limits might be able to be increased further if
it can be proven that the FPUs work reliably with
higher speed and acceleration limits.