switch to assisted datum search if FPU state is not initialized
This is for the case that the FPU was initialized before,
but is in state FPST_UNINITIALIZED. In this case,
use the position from the hardware protection database/LMDB
to instruct the hardware in which direction the beta
arm datum search should take place.

Needs possibly to be revised for the verification software,
which checks the arms separately in the functional tests.