change maximum step difference (step jump between segments)
The value was taken over from the verification parameters.
However this was incorrect, as the parameters in the
verification system refer to larger time intervals.

What was observed was a loss of control (along with
melodic noises) for the beta motor.
The value was halved to 50 steps difference.
At first glance, this seems to work better.