Fix setup of SYNC commands
This fixes that the driver didn't check for the sending
of SYNC configuration data to complete. Because
the socket is opened in non-blocking mode, the
sending configuration data might need to be started
multiple times until it is complete.

If the sending fails, the by far most likely cause
of the error is that there is no socket connection
at all - either because a gateway is not connected,
or it is not working.