raise exception if findDatum was canceled
This raises an exception of findDatum was canceled or
startFindDatum returned with an error. This can happen
if the alpha arm is on the limit switch. The latter can
be caused (a) by an unresolved alpha imit breach or (b)
by an incomplete findDatum operation, where the firmware
exists the findDatum operation while the alpha arm is
still on the limit switch.